+++ R/launch_editor.R
@@ -73,6 +73,9 @@ launch_editor <- function(app_loc,
     stop("Stopping UI Editor. Reason:", app_status$message)
   }
 
+  # Make sure environment will allow features to work properly
+  check_for_url_issues()
+
   app_preview <- AppPreview$new(
     app_loc = app_loc,
     port = shiny_background_port,
@@ -110,6 +113,7 @@ launch_editor <- function(app_loc,
   # open the browser to it for them
   announce_location_of_editor(port, launch_browser)
 
+  # Main server startup - Runs in main process
   httpuv::runServer(
     host = host, port = port,
     app = list(
@@ -140,10 +144,14 @@ launch_editor <- function(app_loc,
             simplifyVector = FALSE
           )
 
+
+
           switch(message$path,
             "APP-PREVIEW-CONNECTED" = {
               app_preview$set_listeners(
                 on_ready = function() {
+                  # Once the background preview app is up and running, we can
+                  # send over the URL to the react app
                   ws$send(
                     build_ws_message(
                       "SHINY_READY",