+++ NEWS.md
@@ -1,4 +1,4 @@
-# shinyuieditor (development version)
+# shinyuieditor (development)
 
 ### Major new features and improvements
 
@@ -7,6 +7,8 @@
 - Popups on hover over the elements panel now provide some context for the functionality of the element. (#74)
 
 ### Bug fixes
+
+- Fixes bug where ui editor would not work on hosted services like RStudio Workbench (#73)
 
 ### Known bugs